requirements.txt
Django==1.9.1
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
celery-with-redis==3.0
django-celery==3.1.17
django-crispy-forms==1.6.0
django-haystack==2.5.dev0
elasticsearch==2.2.0
factory-boy==2.6.0
fake-factory==0.5.3
kombu==3.0.33
pysolr==3.4.0
pytz==2015.7
redis==2.10.5
requests==2.9.1
urllib3==1.14
